requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
pandas
numpy
faker
scikit-learn
joblib
python-dotenv
